+++ R/geom_label_contour.R
@@ -1,0 +1,113 @@
+#' @rdname geom_text_contour
+#' @export
+geom_label_contour <- function(mapping = NULL, data = NULL,
+                       stat = "contour", position = "identity",
+                       ...,
+                       min.size = 10,
+                       skip = 1,
+                       parse = FALSE,
+                       nudge_x = 0,
+                       nudge_y = 0,
+                       label.padding = unit(0.25, "lines"),
+                       label.r = unit(0.15, "lines"),
+                       label.size = 0.25,
+                       na.rm = FALSE,
+                       show.legend = NA,
+                       inherit.aes = TRUE) {
+    if (!missing(nudge_x) || !missing(nudge_y)) {
+        if (!missing(position)) {
+            stop("Specify either `position` or `nudge_x`/`nudge_y`", call. = FALSE)
+        }
+
+        position <- position_nudge(nudge_x, nudge_y)
+    }
+
+    ggplot2::layer(
+        data = data,
+        mapping = mapping,
+        stat = stat,
+        geom = GeomLabelContour,
+        position = position,
+        show.legend = show.legend,
+        inherit.aes = inherit.aes,
+        params = list(
+            min.size = min.size,
+            skip = skip,
+            parse = parse,
+            label.padding = label.padding,
+            label.r = label.r,
+            label.size = label.size,
+            na.rm = na.rm,
+            ...
+        )
+    )
+}
+
+GeomLabelContour <- ggplot2::ggproto("GeomLabelContour", Geom,
+    required_aes = c("x", "y", "label"),
+    default_aes = ggplot2::aes(
+        colour = "black", fill = "white", size = 3.88, angle = 0,
+        hjust = 0.5, vjust = 0.5, alpha = NA, family = "", fontface = 1,
+        lineheight = 1.2
+    ),
+
+    draw_panel = function(self, data, panel_params, coord, parse = FALSE,
+                          na.rm = FALSE,
+                          label.padding = unit(0.25, "lines"),
+                          label.r = unit(0.15, "lines"),
+                          label.size = 0.25, min.size = 20,
+                          skip = 1) {
+        data <- data.table::as.data.table(coord$transform(data, panel_params))
+
+        breaks <- unique(data$level)
+        breaks.keep <- breaks[seq(1, length(breaks), by = skip + 1)]
+        data <- data[level %in% breaks.keep]
+
+        data[, N := .N, by = piece]
+        data <- data[N >= min.size]
+
+        data[, angle := .cont.angle(x, y), by = piece]
+
+        data[, var := minvar(x, y), by = .(piece)]
+        data <- data[var == T][, head(.SD, 1), by = piece]
+        lab <- data$label
+        if (parse) {
+            lab <- parse(text = as.character(lab))
+        }
+
+        if (is.character(data$vjust)) {
+            data$vjust <- ggplot2:::compute_just(data$vjust, data$y)
+        }
+        if (is.character(data$hjust)) {
+            data$hjust <- ggplot2:::compute_just(data$hjust, data$x)
+        }
+
+        grobs <- lapply(1:nrow(data), function(i) {
+            row <- data[i, , drop = FALSE]
+            ggplot2:::labelGrob(lab[i],
+                      x = unit(row$x, "native"),
+                      y = unit(row$y, "native"),
+                      just = c(row$hjust, row$vjust),
+                      padding = label.padding,
+                      r = label.r,
+                      text.gp = grid::gpar(
+                          col = row$colour,
+                          fontsize = row$size * .pt,
+                          fontfamily = row$family,
+                          fontface = row$fontface,
+                          lineheight = row$lineheight
+                      ),
+                      rect.gp = grid::gpar(
+                          col = row$colour,
+                          fill = alpha(row$fill, row$alpha),
+                          lwd = label.size * .pt
+                      )
+            )
+        })
+        class(grobs) <- "gList"
+
+        ggplot2:::ggname("geom_label_contour", grid::grobTree(children = grobs))
+    },
+
+    draw_key = draw_key_label
+)